requirements.txt
streamlit~=1.33.0
fpdf2~=2.7.8